projection sorting into 4 bins from the phase resp sig
having the respiratory signal as a combination of the three chips respiratory signal (resp_sig3c), the projection locations (expiration_proj_loc.m)  sorts in 4 different folders the original projections by means of  the polar histogram of the phase respiratory signal (sortPhaseRespSignal(resp_sig3c')). A .txt is added to the folders with the statistics of the sorting.